Cleaned up uses of generic collections in java (Similar to templates in C++). Java 1.4 didn't support the specializations, but now we use Java 5 which supports these.  Now the sets and lists and maps are specialized on the appropriate types that they contain. This should help the compiler catch problems, and make it more clear what is stored in such variables. Fixed a bug with message send locations in Timeline.

+++ jnt/FFT/ComplexDoubleFFT_Mixed.java
@@ -50,7 +50,7 @@ public class ComplexDoubleFFT_Mixed extends ComplexDoubleFFT{
 
     factors = Factorize.factor(n, available_factors);
 
-    double d_theta = -2.0 * PI / ((double) n);
+    double d_theta = -2.0 * PI / (n);
     int product = 1;
     twiddle = new double[factors.length][][];
     for (int i = 0; i < factors.length; i++) {
