+++ projections/gui/Timeline/UserEventWindow.java
@@ -78,10 +78,10 @@ public class UserEventWindow extends JFrame
 				switch (col) {
 				case 0:  
 					return ((UserEventObject)events_[index_][row]).getName();
-				case 1:  return new FormattedNumber((long)((UserEventObject)events_[index_][row]).BeginTime, format_);
-				case 2:  return new FormattedNumber((long)((UserEventObject)events_[index_][row]).EndTime, format_);
-				case 3:  return new FormattedNumber((long)((UserEventObject)events_[index_][row]).EndTime-
-						(long)((UserEventObject)events_[index_][row]).BeginTime, format_);
+				case 1:  return new FormattedNumber(((UserEventObject)events_[index_][row]).BeginTime, format_);
+				case 2:  return new FormattedNumber(((UserEventObject)events_[index_][row]).EndTime, format_);
+				case 3:  return new FormattedNumber(((UserEventObject)events_[index_][row]).EndTime-
+						((UserEventObject)events_[index_][row]).BeginTime, format_);
 				default: return "ERROR";
 				}
 			}
